### 2025-07-04 #8540 @editorialbot
[PRE REVIEW]: HiGP: A high-performance Python package for Gaussian Processes
**Submitting author:** @huanghua1994 (<a href="http://orcid.org/0000-0003-1060-5639">Hua Huang</a>)
**Repository:** https://github.com/huanghua1994/HiGP
**Branch with paper.md** (empty if default branch): paper
**Version:** 2025.2.17
