### 2025-07-04 #8540 @danielskatz
👋 @huanghua1994 - Thanks for your submission to JOSS. However, your paper is too long. JOSS's guidance is for papers of 250-1000 words, and yours is about 2000. Please shorten your paper. Perhaps some material can be in the repository, and pointed to in the paper?

You can use command `@editorialbot check repository` to check a few things, including the paper length, and the comment `@editorialbot generate pdf` make a new PDF. editorialbot commands need to be the first entry in a new comment.

In addition, you could work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. And for Hensman, please remove the doi field in your .bib file and replace it by a url field with `https://dl.acm.org/doi/10.5555/3023638.3023667`  (This is not a DOI but a handle. ACM's use of `doi` in the URL for their resolver is misleading.)

Please feel free to make changes to your .bib file, then you can use the command `@editorialbot check references` to check again.
