### 2025-07-04 43dbaa0
Shorten the paper and update bib file
