+++ paper/paper.md
@@ -36,117 +36,65 @@ tags:
 
 # Summary
 
-Gaussian Processes (GPs) [@Rasmussen:2005] are flexible, nonparametric Bayesian models widely used for regression and classification tasks due to their ability to capture complex data patterns and provide uncertainty quantification. Traditional GP implementations often face challenges in scalability as data sizes increase. To address these challenges, HiGP, a high-performance Python package, is designed for efficient Gaussian Process regression and classification across datasets of varying sizes. HiGP uses  iterative solvers to access large GP computations. It implements matrix-vector (MatVec) and matrix-matrix (MatMul) multiplication algorithms tailored to kernel matrices by exploiting their hierarchical low-rank structure [@Xing:2019; @Huang:2020; @Cai:2023]. To improve the convergence of iterative methods, HiGP also integrates the Adaptive Factorized Nyström (AFN) preconditioner [@Zhao:2024] which is designed to be robust across a wide range of kernel hyperparameter values used during hyperparameter optimization. HiGP can be used with PyTorch and other Python packages, allowing it to be incorporated into common machine learning and data analysis workflows.
+Gaussian Processes (GPs) [@Rasmussen:2005] are flexible, nonparametric Bayesian models widely used for regression and classification tasks due to their ability to capture complex data patterns and provide uncertainty quantification. Traditional GP implementations often face challenges in scalability as data sizes increase. To address these challenges, HiGP, a high-performance Python package, is designed for efficient Gaussian Process regression and classification across datasets of varying sizes. HiGP uses iterative solvers to access large GP computations. It implements matrix-vector (MatVec) and matrix-matrix (MatMul) multiplication algorithms tailored to kernel matrices by exploiting their hierarchical low-rank structure [@Xing:2019; @Huang:2020; @Cai:2023]. To improve the convergence of iterative methods, HiGP also integrates the Adaptive Factorized Nyström (AFN) preconditioner [@Zhao:2024] which is designed to be robust across a wide range of kernel hyperparameter values used during hyperparameter optimization. HiGP can be used with PyTorch and other Python packages, allowing it to be incorporated into common machine learning and data analysis workflows.
 
 # Gaussian Processes
 
-For training points $\mathbf{X} \in \mathbb{R}^{n \times d}$, a noisy training observation set $\mathbf{y} \in \mathbb{R}^{n}$, and testing points $\mathbf{X}_\ast \in \mathbb{R}^{m \times d}$, a standard GP model assumes that the noise-free testing observations $\mathbf{y}_\ast \in \mathbb{R}^{m}$ follow the joint distribution:
-$$
-\begin{bmatrix}
-\mathbf{y} \\
-\mathbf{y}_\ast
-\end{bmatrix}
-\sim
-\mathcal{N} 
-\begin{pmatrix}
-\mathbf{0},
-f^2
-\begin{bmatrix}
-    \kappa(\mathbf{X}, \mathbf{X}) + s \mathbf{I} & \kappa(\mathbf{X}, \mathbf{X}_{\ast}) \\ 
-    \kappa(\mathbf{X}_{\ast}, \mathbf{X}) & \kappa(\mathbf{X}_{\ast}, \mathbf{X}_{\ast})
-\end{bmatrix}
-\end{pmatrix}. \tag{1}
-$$
[…]
