+++ paper/paper.bib
@@ -38,7 +38,7 @@
 	publisher    = {AUAI Press},
     pages        = {282--290},
     location     = {Bellevue, WA},
-	doi          = {10.5555/3023638.3023667}
+	url          = {https://dl.acm.org/doi/10.5555/3023638.3023667}
 }
 @misc{Wilson:2015,
 	title        = {Thoughts on Massively Scalable {Gaussian} Processes},
@@ -80,7 +80,8 @@
 	year         = 2021,
 	booktitle    = {Symposium on Simplicity in Algorithms (SOSA)},
 	publisher    = {SIAM},
-	pages        = {142--155}
+	pages        = {142--155},
+	doi          = {10.1137/1.9781611976496.16}
 }
 @article{Ubaru:2017,
 	title        = {Fast Estimation of $tr(f(A))$ via Stochastic Lanczos Quadrature},
@@ -120,7 +121,8 @@
 	year         = 2020,
 	journal      = {SIAM Journal on Matrix Analysis and Applications},
 	volume       = 41,
-	pages        = {221--243}
+	pages        = {221--243},
+	doi          = {10.1137/19m1258700}
 }
 @article{Huang:2020,
 	title        = {{H2Pack}: High-performance \textit{{H}} $^{\textrm{2}}$ Matrix Package for Kernel Matrices Using the Proxy Point Method},
@@ -139,7 +141,8 @@
 	journal      = {SIAM Journal on Scientific Computing},
 	volume       = {46},
 	number       = {2},
-	pages        = {S24-S50}
+	pages        = {S24-S50},
+	doi          = {10.1137/22m1500848}
 }
 % AFN related
 @article{Zhao:2024,
