### 2025-07-04 #8540 @huanghua1994
Hi @danielskatz, I shortened our paper and updated some bib items. The current wordcount for `paper.md` is 1098, but the output PDF does not really have so many words (MS Word says it has ~920 words). Is that okay? If not, we can try to further shorten the paper. Thanks!
