### 2025-07-19 #8540 @jGaboardi
👋 @amontoison, @Naeemkh, @davidogara, @martinmodrak - would one or more of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
