### 2025-08-03 #8621 @Edenhofer
# Paper Review

I think this contribution is moving research on GPs forward in an important direction. GP models are widely considered to be very flexible but their major disadvantage is speed. Hence, more performant methods and/or implementations are crucial and this contribution promises exactly that.

The paper reads well and is very suitable for JOSS. However, for a contribution focused on performance, I'm missing a clear benchmark showing that shows how it improves upon existing methods. Furthermore, the literature review in the paper is lacking and the three major advantages in favor of HIGP are not as clear cut as outlined.

## Major
* The main advantage of the work seems to be performance. Please provide a **benchmark comparing the GP** evaluation plus gradient computation time (or another suitable metric) to a comparable state-of-the-art model!
* l41: "and other packages" seems a bit too brief (plus, tensorflow is a bit dated by now)
* l46f: for spatial 2D and 3D data, FFTs are usually an extremely efficient way to represent certain kernels see, e.g. KISS-GP. Since you mention these structures explicitly, please provide a brief description of the advantage that alternative representations yield compared to commonly used FFTs. In the case of implicit representations of the kernel matrix, the matrices are instantiated only partially and usually not stored.
* l52f: I'm not sure whether iterative methods plus pre-conditioners really set HiGP apart. AFAIK, many GP tools use iterative solvers with some kind of pre-conditioner including GPyTorch, the tool you cite. Even the ones that don't, tend to have a reason for not using one.
* l58f: Iterative methods are usually not naively auto-diffed AFAIK (naively as in iterations being unrolled)! For example, JAX defines custom derivates for most iterative solvers and utilizes the implicit function theorem for retrieving gradients. Furthermore, using "naked" AD would not be safe as at least some Krylov subspace methods are prone to diverging gradients. IMO, you should focus on the gain in performance (with benchmarks) of having hand-coded gradients, not the efficiency of AD for iterative solvers.

## Minor
* l22ff: First sentence is very hard to read.
* l47: Missing brackets around citation.
* l54f: Please provide a citation for AFN significantly improving things.
* l57f: i) Please do not put brackets in brackets. ii) Please link equation.
* l65f: The sentence is very hard to read. The quantity of "and"s is confusing.
* l78f: I think the end of the sentence got mangled.
* l81: I assume the stateless-ness does not expand to GPRModel/GPCModel as they track their parameters. If so, please say so in the paper.

## Miscellaneous

Could you please elaborate on the nature of the shared authorship and the contributions of each author? The [git commit history](https://github.com/huanghua1994/HiGP/graphs/contributors) (admittedly only a crude proxy for the true contributions) shows @huanghua1994 as having contributed the vast majority of code.
