### 2025-08-04 huanghua1994/HiGP#2 @Edenhofer
Please host docs for easier access and searching
Related to https://github.com/openjournals/joss-reviews/issues/8621#issuecomment-3094725559 .
