### 2025-08-04 huanghua1994/HiGP#3 @Edenhofer
Please populate page on [PyPI](https://pypi.org/project/higp/)
Related to https://github.com/openjournals/joss-reviews/issues/8621#issuecomment-3094725559 .
