### 2025-08-04 huanghua1994/HiGP#4 @Edenhofer
If possible, please build wheels for more operating systems (e.g. mac and win)
Since, this repository contains C++ code, it might be convenient to pre-build it for multiple architectures, e.g. using pypa/cibuildwheel.

Related to https://github.com/openjournals/joss-reviews/issues/8621#issuecomment-3094725559 .
