### 2025-08-04 huanghua1994/HiGP#5 @Edenhofer
Contribution Guidelines
Please add contribution guidelines for devs, filing issues, and general support.

Related to https://github.com/openjournals/joss-reviews/issues/8621#issuecomment-3094725559 .
