### 2025-08-10 75f92fe
Minor modifications following @Edenhofer 's review comments
