+++ paper/paper.md
@@ -53,31 +53,33 @@ $$
 \text{tr}{\left( \widehat{\mathbf{K}}^{-1} \frac{\partial \widehat{\mathbf{K}}}{\partial \theta} \right)}\right),
 \quad \theta \in \Theta.
 $$
-Using preconditioned iterative methods with preconditioner $\mathbf{M} \approx \widehat{\mathbf{K}}$ is a common option [@Aune:2014, @Zhang:2024, @Hensman:2013, @Wilson:2015, @Pleiss:2018, @Wenger:2022, @Chen:2023]. In this approach, $\widehat{\mathbf{K}}^{-1}\mathbf{y}$ is approximated via the preconditioned conjugate gradient (PCG) method [@Saad:2003]. To handle the logarithmic determinant and trace terms, they are first rewritten as 
+Using preconditioned iterative methods with preconditioner $\mathbf{M} \approx \widehat{\mathbf{K}}$ is a common option [@Aune:2014; @Zhang:2024; @Hensman:2013; @Wilson:2015; @Pleiss:2018; @Wenger:2022; @Chen:2023]. In this approach, $\widehat{\mathbf{K}}^{-1}\mathbf{y}$ is approximated via the preconditioned conjugate gradient (PCG) method [@Saad:2003]. To handle the logarithmic determinant and trace terms, they are first rewritten as
 $$
 \log|\widehat{\mathbf{K}}|=\log|\mathbf{M}| + \log|\mathbf{M}^{-1/2}\widehat{\mathbf{K}}\mathbf{M}^{-1/2}|, \tag{1}
 $$
-$$
+
+\begin{equation}
+\label{eq:trace}
 \text{tr}{(\widehat{\mathbf{K}}^{-1} \frac{\partial \widehat{\mathbf{K}}}{\partial \theta})}={\text{tr}}\left({{\mathbf{M}}^{-1}}\frac{\partial {\mathbf{M}}}{\partial \theta}\right)+
  {\text{tr}}\left({\widehat{\mathbf{K}}^{-1}}\frac{\partial \widehat{\mathbf{K}}}{\partial \theta}-{{\mathbf{M}}^{-1}}\frac{\partial {\mathbf{M}}}{\partial \theta}\right). \tag{2}
- $$
- The second component of each new expression is then estimated using the stochastic Lanczos quadrature [@Ubaru:2017] and the Hutchinson estimator [@Hutchinson:1989, Meyer:2021], respectively. The AFN preconditioner is used in HiGP to speed up the optimization process.
+\end{equation}
+The second component of each new expression is then estimated using the stochastic Lanczos quadrature [@Ubaru:2017] and the Hutchinson estimator [@Hutchinson:1989; Meyer:2021], respectively. The AFN preconditioner is used in HiGP to speed up the optimization process.
 
 # Statement of Need
 
-Multiple GP packages have been released in recent years to utilize modern hardware and handle larger datasets, including GPyTorch [@Gardner:2018], GPflow [@GPflow:2017, @GPflow:2020], GPy [@GPy:2014], and other packages. The motivation for HiGP's development is to utilize new numerical algorithms and parallel computing techniques to reduce computational complexity and to improve the computation efficiency of iterative methods in GP model training. Compared to existing packages, HiGP has three main advantages and contributions.
+Multiple GP packages have been released in recent years to utilize modern hardware and handle larger datasets, including GPyTorch [@Gardner:2018], GPflow [@GPflow:2017; @GPflow:2020], GPy [@GPy:2014], and other packages. The motivation for HiGP's development is to utilize new numerical algorithms and parallel computing techniques to reduce computational complexity and to improve the computation efficiency of iterative methods in GP model training. Compared to existing packages, HiGP has three main advantages and contributions.
 
[…]
