### 2025-08-16 huanghua1994/HiGP#7 @Naeemkh
Fix issue with pip installation
I took the following steps and could not install the package from pip:

```bash
conda create -n higp_env python=3.12
conda activate higp_env
```
Then

```bash
pip install higp
```

Error message:

```
ERROR: Could not find a version that satisfies the requirement higp (from versions: none)
ERROR: No matching distribution found for higp
```

This issue is related to this submission: https://github.com/openjournals/joss-reviews/issues/8621#issuecomment-3094724362
