### 2025-08-16 huanghua1994/HiGP#8 @Naeemkh
Add instructions for building and installing from source
Please add instructions for building the package from source.

This is related to this submission: https://github.com/openjournals/joss-reviews/issues/8621#issuecomment-3094724362
