### 2025-08-16 huanghua1994/HiGP#8 @huanghua1994
Hi, please refer to [4.2 Building HiGP on local machine](https://github.com/huanghua1994/HiGP/blob/main/docs/4-Developer-information.md#42-building-higp-on-local-machine) in the documentation. Currently we have only tested HiGP on Linux, cannot compile on macOS yet (see the discussion in #6).
