### 2025-08-16 huanghua1994/HiGP#7 @huanghua1994
Hi, are you using macOS? If yes, would you please try it on a Linux machine? The current build script is designed for linking optimized BLAS/LAPACK on Linux; it does not work on macOS. Therefore, we do not have a macOS build at the moment. We will try to solve this in the future.
