### 2025-08-20 24f7a0a
Add experimental Apple Silicon build support
