+++ cpp-src/cpu_linalg.h
@@ -1,7 +1,14 @@
 #ifndef __CPU_LINALG_H__
 #define __CPU_LINALG_H__
 
-#if defined(USE_MKL) || defined(USE_OPENBLAS_LP64)
+#if defined(USE_ACCELERATE_LP64)
+
+#ifndef ACCELERATE_NEW_LAPACK
+#define ACCELERATE_NEW_LAPACK
+#endif
+#include <Accelerate/Accelerate.h>
+#include "cpu_linalg_lp64.h"
+#elif defined(USE_MKL) || defined(USE_OPENBLAS_LP64)
 #include "cpu_linalg_lp64.h"
 #endif
 
