### 2025-08-21 92d7c73
Add macOS wheel build script
