### 2025-08-23 huanghua1994/HiGP#9 @davidogara
Add tests for Python package
Part of https://github.com/openjournals/joss-reviews/issues/8621

While there are an extensive set of c++ tests in this package, it would be great to have a set of Python tests that are part of the package. This would be helpful for two reasons:
1. While I understand the workhorse functions of the package are written in c++, the Python code appears to be untested
2. For users who want to build from source, having a set of Python tests would be great for checking proper installation
3. Having a set of Python tests would also allow for complete "end to end" testing of the package i.e. that training a GP regression on a standard toy problem yields the same (or approximately the same) results

Therefore, I suggest that the authors include a few Python tests using the `pytest` framework
