### 2025-08-23 huanghua1994/HiGP#10 @davidogara
Remove Matlab tests
Part of https://github.com/openjournals/joss-reviews/issues/8621

I suggest removing the Matlab tests, for example, https://github.com/huanghua1994/HiGP/blob/4cf297cc79a76ab2a92d17ba7ae3ce681e620a84/cpp-tests/check_test_exact_gpc.m

It isn't clear if these are still used, or are legacy from an earlier version of the project, but since this project is being submitted to JOSS and Matlab is not open source, I recommend the authors either remove the tests, archive them on a separate branch, or add some documentation that those tests are no longer active.
