### 2025-08-23 huanghua1994/HiGP#11 @davidogara
Clarify whether anisoptropic kernels are supported
Part of https://github.com/openjournals/joss-reviews/issues/8621

While reviewing the examples, I did not see an example where a user could learn a lengthscale parameter for each input dimension. Can the authors please comment on whether it is possible, and if so, include an example notebook? Since these are supported in both sklearn and GPyTorch, it would be good to know whether this is supported.
