### 2025-08-23 huanghua1994/HiGP#12 @Edenhofer
Error in "Basic Usage of HiGP"
Running the code described at [Basic Usage of HiGP](https://github.com/huanghua1994/HiGP/blob/main/docs/1-Basic-usage-of-HiGP.md) leads to the following error:

```txt
Cell In[8], line 1
----> 1 pred = higp.gpr_prediction(data_train=train_x,
      2                            label_train=train_y,
      3                            data_prediction=test_x,
      4                            kernel_type=higp.GaussianKernel,
      5                            pyparams=model.get_params())

File ~/Projects/HiGP/.venv/lib/python3.11/site-packages/higp/modules.py:31, in gpr_prediction(*args, **kwargs)
     30 def gpr_prediction(*args, **kwargs):
---> 31     pred = higp_cext.gpr_prediction_(*args, **kwargs)
     32     results = SimpleNamespace()
     33     results.prediction_mean = pred[0]

ValueError: Error in the input argument!
```

I don't know how to ad hoc debug the error message further as it is somewhat nondescript. Could you please fix the documentation and ideally update the the error message to contain more information?

P.S. `fig` is not defined in the demo but used as a variable. I'd suggest to write normal python scripts, tests these automatically, and translate them to ipynb notebooks via jupytext to avoid errors like these.

Related to https://github.com/openjournals/joss-reviews/issues/8621#issuecomment-3094725559 .
