### 2025-08-24 e6ec68c
Update parameter name in basic usage docs
