### 2025-08-25 e8dec6e
add basic pytests
