### 2025-08-25 #8621 @davidogara
## Paper Review

This work presents higp, a Python/C++ package geared towards high performance Gaussian process regression and classification. The package’s modules and example notebooks appear to be in good shape. I also appreciate the authors’ efforts to implement much of the codebase in C++, as a major drawback to working with GPs is long training time. By implementing cutting edge methods (like AFN) in a high-performance language (C++) this package may go a long way towards reducing GP training costs.

However, as other reviewers have mentioned, I have some feedback on the claims of high performance.  If the authors test the package’s performance on a suite of problems, I suggest anything from the virtual library of simulation experiments (https://www.sfu.ca/~ssurjano/) , particularly the optimization tab, which has many classic optimization problems that GPs are often tested on (like the Rosenbrock function). In addition, the authors make some claims related to memory efficiency (lines 45-51 in the draft), and like the claims of efficient training time, the paper would also benefit from some benchmarks related to memory efficiency compared to another candidate Python library (probably GPyTorch).

Additionally, in the Statement of Need, it would be great if the authors could address how difficult it would be to implement the core components of HiGP in another Python library such as GPyTorch, which I think of as the standard for GPs in Python. For example, to implement AFN: my cursory understanding is that this would require custom LinearOperator routines (i.e. the custom matrix routines that underly GPyTorch) and thus it is not clear to me how challenging those would be to implement. Any insight the authors have on this point would be welcome.

The remaining piece of text I would recommend is a brief mention of next steps. Notably, I think supporting anisotropic kernels would provide additional benefit for users since they are supported out of the box in packages like sklearn, GPyTorch, and GPFlow.

Finally, I had difficultly installing the package on an Intel-based Mac, so I would also recommend the authors provide support for those (since they are still used widely to my knowledge), though I was able to compile the package using Docker (running an ubuntu image) to test its functionality.
