### 2025-08-27 a8b2ed4
Add Apple Silicon support with Accelerate Framework and OpenMP integration
