### 2025-08-27 9fe0144
Reorganize demos: move notebooks to py-demos and improve seed parameter support
