+++ py-interface/higp_cext.h
@@ -648,6 +648,7 @@ static char HiGP_Cext_gpr_prediction_help[] =
     "    afn_lfil (0)   : The fill-level of the Schur complement of the AFN preconditioner.\n"
     "    niter (50)     : The number of iterations for PCG in the precondition.\n"
     "    tol (1e-6)     : The tolerance for PCG in the precondition.\n"
+    "    seed (-1)      : Random number generator seed, -1 for system default number.\n"
     "Output:\n"
     "    prediction: tuple with two elements.\n"
     "        prediction[0] : Prediction mean value (NumPy array, length N).\n"
@@ -673,6 +674,7 @@ static char HiGP_Cext_gpc_prediction_help[] =
     "    afn_lfil (0)   : The fill-level of the Schur complement of the AFN preconditioner.\n"
     "    niter (50)     : The number of iterations for PCG in the precondition.\n"
     "    tol (1e-6)     : The tolerance for PCG in the precondition.\n"
+    "    seed (-1)      : Random number generator seed, -1 for system default number.\n"
     "Output:\n"
     "   prediction: tuple with three elements.\n"
     "       prediction[0] : Prediction label (NumPy array, length N).\n"
