### 2025-08-27 40969c7
Add virtual library of simulation experiments for GP performance testing
