+++ py-demos/virtual_library/utils/__init__.py
@@ -1,0 +1,8 @@
+from .test_functions import (
+    rosenbrock,
+    rastrigin,
+    branin,
+    get_function_info,
+)
+from .data_generation import latin_hypercube_sampling, generate_training_data, add_noise
+from .metrics import compute_rmse, compute_r2, compute_nll
