+++ py-demos/virtual_library/runners/__init__.py
@@ -1,0 +1,2 @@
+from .higp_runner import run_higp
+from .gpytorch_runner import run_gpytorch
