### 2025-08-28 a21ac98
Add data normalization and kernel type support to VL tests
