### 2025-08-29 52d2276
Clean up VL tests: update config and output;suppress C++ output and warnings
