+++ py-demos/virtual_library/test_vl.py
@@ -4,6 +4,7 @@ import sys
 import json
 import argparse
 import time
+import numpy as np
 from datetime import datetime
 
 sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
@@ -45,7 +46,6 @@ def run_single_experiment(
         )
         return None
 
-    # Data generation
     func = func_info["func"]
     bounds = func_info["bounds"]
     data_seed = seed + repeat_id * 1000
@@ -53,7 +53,6 @@ def run_single_experiment(
         func, n_train, n_test, dim, bounds, noise_level=noise_level, seed=data_seed
     )
 
-    # Apply Z-score normalization
     print(
         f"  Before normalization - Y range: [{train_y.min():.2f}, {train_y.max():.2f}], mean: {train_y.mean():.2f}"
     )
@@ -64,7 +63,6 @@ def run_single_experiment(
         f"  After normalization - Y range: [{train_y.min():.2f}, {train_y.max():.2f}], mean: {train_y.mean():.2f}"
     )
 
-    # Run HiGP
     print(f"  Running HiGP...")
     higp_params_with_kernel = {**higp_params, "kernel": kernel}
     higp_results = run_higp(
@@ -85,7 +83,6 @@ def run_single_experiment(
     higp_r2 = compute_r2(test_y, higp_results["y_pred"])
     higp_nll = compute_nll(test_y, higp_results["y_pred"], higp_results["y_std"])
 
-    # Run GPyTorch (if not skipped)
     if not skip_gpytorch:
         print(f"  Running GPyTorch...")
         gpytorch_params_with_kernel = {**gpytorch_params, "kernel": kernel}
@@ -117,7 +114,6 @@ def run_single_experiment(
         gpytorch_r2 = None
         gpytorch_nll = None
 
-    # Save results
     result = {
         "function": func_name,
         "dim": dim,
@@ -136,6 +132,7 @@ def run_single_experiment(
             "inference_time": higp_results["inference_time"],
             "init_loss": higp_results["init_loss"],
             "final_loss": higp_results["final_loss"],
+            "actual_config": higp_results.get("actual_config", {}),
         },
     }
 
@@ -151,9 +148,9 @@ def run_single_experiment(
             "inference_time": gpytorch_results["inference_time"],
             "init_loss": gpytorch_results["init_loss"],
             "final_loss": gpytorch_results["final_loss"],
+            "cg_warning": gpytorch_results.get("cg_warning", False),
         }
 
-    # Prepare detailed results with all data
     detailed_result = result.copy()
     detailed_result["higp"]["loss_history"] = higp_results["loss_history"]
     detailed_result["higp"]["y_pred"] = higp_results["y_pred"].tolist()
@@ -176,7 +173,6 @@ def run_single_experiment(
 
     detailed_result["test_y"] = test_y.tolist()
 
-    # Summary result should only contain core metrics
     summary_result = {
         "function": func_name,
         "dim": dim,
@@ -195,6 +191,7 @@ def run_single_experiment(
             "inference_time": higp_results["inference_time"],
             "init_loss": higp_results["init_loss"],
             "final_loss": higp_results["final_loss"],
+            "actual_config": higp_results.get("actual_config", {}),
         },
     }
 
@@ -230,20 +227,16 @@ def main():
     )
     args = parser.parse_args()
 
-    # Load configuration
     with open(args.config, "r") as f:
         config = json.load(f)
 
-    # Create output directory
     os.makedirs(args.output_dir, exist_ok=True)
     timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
     exp_name = config.get("name", "experiment")
 
-    # Results storage
     all_results = []
[…]
