### 2025-08-29 5d4d416
add UCI tests
