+++ py-demos/uci_benchmark/utils/__init__.py
@@ -1,0 +1,17 @@
+from .datasets import load_uci_dataset, BikeDataset, Road3DDataset
+from .data_processing import (
+    prepare_for_higp,
+    train_test_normalize,
+)
+from .metrics import compute_rmse, compute_r2, compute_nll
+
+__all__ = [
+    "load_uci_dataset",
+    "BikeDataset",
+    "Road3DDataset",
+    "prepare_for_higp",
+    "train_test_normalize",
+    "compute_rmse",
+    "compute_r2",
+    "compute_nll",
+]
