+++ py-demos/uci_benchmark/test_uci.py
@@ -1,0 +1,348 @@
+#!/usr/bin/env python
+"""Main test script for UCI benchmark."""
+
+import os
+import sys
+import json
+import argparse
+import numpy as np
+from datetime import datetime
+from collections import defaultdict
+
+sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
+
+from uci_benchmark.utils import (
+    load_uci_dataset,
+    train_test_normalize,
+    prepare_for_higp,
+    compute_rmse,
+    compute_r2,
+    compute_nll,
+)
+from uci_benchmark.runners import run_higp, run_gpytorch
+
+
+def run_single_experiment(
+    dataset_name,
+    n_train,
+    n_test,
+    higp_params,
+    gpytorch_params,
+    kernel,
+    dtype_str,
+    seed,
+    repeat_id,
+    skip_gpytorch=False,
+):
+    """Run a single experiment comparing HiGP and GPyTorch on UCI data."""
+
+    print(f"  Loading {dataset_name} dataset...")
+
+    train_x, train_y, test_x, test_y = load_uci_dataset(
+        dataset_name,
+        n_train=n_train,
+        n_test=n_test,
+        seed=seed + repeat_id * 1000,
+    )
+
+    train_x, train_y, test_x, test_y = train_test_normalize(
+        train_x, train_y, test_x, test_y
+    )
+
+    np_dtype = np.float64 if dtype_str == "float64" else np.float32
+    train_x_higp, train_y_higp = prepare_for_higp(train_x, train_y, dtype=np_dtype)
+    test_x_higp, test_y_higp = prepare_for_higp(test_x, test_y, dtype=np_dtype)
+
+    print(f"    Data loaded: {n_train} train, {n_test} test samples")
+    print(f"    Features: {train_x.shape[1]}")
+
+    print(f"  Running HiGP...")
+    higp_params_with_kernel = {**higp_params, "kernel": kernel}
+    higp_results = run_higp(
+        train_x_higp,
+        train_y_higp,
+        test_x_higp,
+        test_y_higp,
+        higp_params_with_kernel,
+        dataset_name=dataset_name,
+        dtype_str=dtype_str,
+        seed=seed,
+    )
+
+    higp_init_rmse = compute_rmse(test_y_higp, higp_results["init_y_pred"])
+    higp_init_r2 = compute_r2(test_y_higp, higp_results["init_y_pred"])
+    higp_init_nll = compute_nll(
+        test_y_higp, higp_results["init_y_pred"], higp_results["init_y_std"]
+    )
+
+    print(f"    Initial RMSE: {higp_init_rmse:.6f}")
+
+    higp_rmse = compute_rmse(test_y_higp, higp_results["y_pred"])
+    higp_r2 = compute_r2(test_y_higp, higp_results["y_pred"])
+    higp_nll = compute_nll(test_y_higp, higp_results["y_pred"], higp_results["y_std"])
+
+    print(f"    Final RMSE: {higp_rmse:.6f}")
+    print(f"    Training time: {higp_results['training_time']:.2f}s")
+
+    if not skip_gpytorch:
+        print(f"  Running GPyTorch...")
+        gpytorch_params_with_kernel = {**gpytorch_params, "kernel": kernel}
+        gpytorch_results = run_gpytorch(
+            train_x_higp,
+            train_y_higp,
+            test_x_higp,
+            test_y_higp,
+            gpytorch_params_with_kernel,
+            dataset_name=dataset_name,
+            dtype_str=dtype_str,
+            seed=seed,
+        )
+
[…]
