### 2025-09-03 7e6b5c0
Update terminal warning suppress and results for tests
