+++ py-demos/virtual_library/utils/data_processing.py
@@ -1,7 +1,7 @@
 """Data processing utilities for Virtual Library."""
 
 import numpy as np
-from typing import Tuple
+from typing import Tuple, Optional
 
 
 def train_test_normalize(
@@ -23,7 +23,7 @@ def train_test_normalize(
         Normalized train_x, train_y, test_x, test_y using training set statistics
     """
     mean_x = train_x.mean(axis=0, keepdims=True)
-    std_x = (train_x.std(axis=0, keepdims=True) + 1e-6)
+    std_x = train_x.std(axis=0, keepdims=True) + 1e-6
     train_x_norm = (train_x - mean_x) / std_x
     test_x_norm = (test_x - mean_x) / std_x
     mean_y = train_y.mean()
@@ -31,3 +31,18 @@ def train_test_normalize(
     train_y_norm = (train_y - mean_y) / std_y
     test_y_norm = (test_y - mean_y) / std_y
     return train_x_norm, train_y_norm, test_x_norm, test_y_norm
+
+
+def prepare_for_higp(
+    X: np.ndarray, y: np.ndarray, dtype: Optional[np.dtype] = None
+) -> Tuple[np.ndarray, np.ndarray]:
+    """Prepare data for HiGP format (transpose X and make contiguous)."""
+    if dtype is not None:
+        X = X.astype(dtype)
+        y = y.astype(dtype)
+
+    # Transpose and make contiguous
+    X_higp = np.ascontiguousarray(X.T)
+    y_higp = np.ascontiguousarray(y)
+
+    return X_higp, y_higp
