### 2025-09-06 8942631
Update matern kernel tests. Clean up and unify test framework.
