+++ py-demos/virtual_library/runners/__init__.py
@@ -1,2 +1,4 @@
 from .higp_runner import run_higp
 from .gpytorch_runner import run_gpytorch
+
+__all__ = ["run_higp", "run_gpytorch"]
