+++ py-demos/uci_benchmark/runners/higp_runner.py
@@ -1,84 +1,14 @@
-"""HiGP runner for UCI benchmark."""
-
-import os
-import sys
 import time
-import gc
-import ctypes
-import tempfile
-import stat
-from contextlib import contextmanager
 import numpy as np
 import torch
-from ..utils import suppress_output
-
-sys.path.insert(
-    0, os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
-)
-
 import higp
+import gc
+import os
+from ..utils import suppress_output, capture_output, prepare_for_higp
 
 
-@contextmanager
-def capture_output():
-    """Capture stdout/stderr and return the text."""
-    try:
-        libc = ctypes.CDLL(None)
-        libc.fflush(None)
-    except (OSError, AttributeError):
-        pass
-
-    old_stdout_fd = os.dup(1)
-    old_stderr_fd = os.dup(2)
-    tmp_fd = -1
-    tmp_filename = None
-
-    try:
-        tmp_fd, tmp_filename = tempfile.mkstemp(prefix="higp_capture_", suffix=".txt")
-        os.chmod(tmp_filename, stat.S_IRUSR | stat.S_IWUSR)
-
-        sys.stdout.flush()
-        sys.stderr.flush()
-
-        os.dup2(tmp_fd, 1)
-        os.dup2(tmp_fd, 2)
-        os.close(tmp_fd)
-        tmp_fd = -1
-
-        yield tmp_filename
-
-    finally:
-        try:
-            libc = ctypes.CDLL(None)
-            libc.fflush(None)
-        except (OSError, AttributeError):
-            pass
-
-        os.dup2(old_stdout_fd, 1)
-        os.dup2(old_stderr_fd, 2)
-        os.close(old_stdout_fd)
-        os.close(old_stderr_fd)
-
-        if tmp_fd >= 0:
-            try:
-                os.close(tmp_fd)
-            except OSError:
-                pass
-
-        sys.stdout.flush()
-        sys.stderr.flush()
-
-
-def run_higp(
-    train_x,
-    train_y,
-    test_x,
-    test_y,
-    params,
-    dtype_str="float32",
-    seed=42,
-):
-    """Run HiGP on UCI dataset."""
+def run_higp(train_x, train_y, test_x, test_y, params, dtype_str="float32", seed=42):
+    """Run HiGP on the given data"""
     if dtype_str == "float64":
         torch.set_default_dtype(torch.float64)
         np_dtype = np.float64
@@ -91,12 +21,8 @@ def run_higp(
     torch.manual_seed(seed)
     np.random.seed(seed)
 
-    mvtype = params.get("mvtype", 0)
-
-    train_x = train_x.astype(np_dtype)
-    train_y = train_y.astype(np_dtype)
-    test_x = test_x.astype(np_dtype)
-    test_y = test_y.astype(np_dtype)
+    train_x_higp, train_y_higp = prepare_for_higp(train_x, train_y, dtype=np_dtype)
+    test_x_higp, _ = prepare_for_higp(test_x, np.array([]), dtype=np_dtype)
 
     kernel_map = {
         "gaussian": higp.GaussianKernel,
@@ -112,14 +38,12 @@ def run_higp(
         "kernel": params.get("kernel", "gaussian").lower(),
     }
 
-    t_train_start = time.perf_counter()
-
     with capture_output() as tmp_file:
         gprproblem = higp.gprproblem.setup(
-            data=train_x,
-            label=train_y,
+            data=train_x_higp,
+            label=train_y_higp,
             kernel_type=kernel_type,
-            mvtype=mvtype,
+            mvtype=params.get("mvtype", 0),
[…]
