+++ py-demos/uci_benchmark/runners/gpytorch_runner.py
@@ -1,5 +1,3 @@
-"""GPyTorch runner for UCI benchmark."""
-
 import time
 import warnings
 import numpy as np
@@ -9,8 +7,6 @@ from ..utils import suppress_output
 
 
 class ExactGPModel(gpytorch.models.ExactGP):
-    """Standard ExactGP model for UCI datasets."""
-
     def __init__(self, train_x, train_y, likelihood, kernel_type="gaussian"):
         super().__init__(train_x, train_y, likelihood)
         self.mean_module = gpytorch.means.ConstantMean()
@@ -31,15 +27,9 @@ class ExactGPModel(gpytorch.models.ExactGP):
 
 
 def run_gpytorch(
-    train_x,
-    train_y,
-    test_x,
-    test_y,
-    params,
-    dtype_str="float32",
-    seed=42,
+    train_x, train_y, test_x, test_y, params, dtype_str="float32", seed=42
 ):
-    """Run GPyTorch on UCI dataset."""
+    """Run GPyTorch on the given data"""
     if dtype_str == "float64":
         torch.set_default_dtype(torch.float64)
         torch_dtype = torch.float64
@@ -50,31 +40,31 @@ def run_gpytorch(
     torch.manual_seed(seed)
     np.random.seed(seed)
 
-    train_x_gp = torch.tensor(train_x.T, dtype=torch_dtype)
-    train_y_gp = torch.tensor(train_y, dtype=torch_dtype)
-    test_x_gp = torch.tensor(test_x.T, dtype=torch_dtype)
-
-    t_train_start = time.perf_counter()
+    train_x_torch = torch.from_numpy(train_x).to(torch_dtype)
+    train_y_torch = torch.from_numpy(train_y).to(torch_dtype)
+    test_x_torch = torch.from_numpy(test_x).to(torch_dtype)
 
     kernel_type = params.get("kernel", "gaussian").lower()
 
     likelihood = gpytorch.likelihoods.GaussianLikelihood()
-    model = ExactGPModel(train_x_gp, train_y_gp, likelihood, kernel_type)
+    model = ExactGPModel(train_x_torch, train_y_torch, likelihood, kernel_type)
 
     model.eval()
     likelihood.eval()
-    with torch.no_grad(), warnings.catch_warnings(), gpytorch.settings.fast_pred_var(), gpytorch.settings.max_preconditioner_size(
-        params.get("pred_max_preconditioner_size", 100)
-    ), gpytorch.settings.max_cg_iterations(
-        params.get("pred_cg_niter", 1000)
+    # Initial predictions with default GPyTorch CG settings
+    with suppress_output(), torch.no_grad(), gpytorch.settings.fast_pred_var(), gpytorch.settings.max_cg_iterations(
+        params.get("pred_cg_niter", params.get("cg_iters", 1000))
     ), gpytorch.settings.max_lanczos_quadrature_iterations(
-        params.get("pred_cg_niter", 1000)
+        params.get("pred_cg_niter", params.get("cg_iters", 1000))
+    ), gpytorch.settings.max_preconditioner_size(
+        params.get(
+            "pred_max_preconditioner_size", params.get("max_preconditioner_size", 100)
+        )
     ):
-        warnings.simplefilter("ignore")
-        init_output = model(test_x_gp)
-        init_preds = likelihood(init_output)
-        init_y_pred = init_preds.mean.numpy()
-        init_y_std = init_preds.stddev.numpy()
+        init_predictions = likelihood(model(test_x_torch))
+        init_y_pred = init_predictions.mean.numpy()
+        init_y_var = init_predictions.variance.numpy()
+        init_y_std = np.sqrt(init_y_var)
 
[…]
