+++ py-demos/virtual_library/utils/common.py
@@ -1,4 +1,4 @@
-"""Common utilities for Virtual Library."""
+"""Common utilities."""
 
 import os
 import sys
