### 2025-09-08 bbeb69d
Add arxiv link to README.md
