### 2025-09-09 #8621 @huanghua1994
Response to @Edenhofer

> 1. The literature review in the paper is lacking and the three major advantages in favor of HIGP are not as clear cut as outlined.

**Response**: We had literature review in the initial submission of the paper, but it was removed due to length limitation (the edit history is available [here](https://github.com/huanghua1994/HiGP/commit/43dbaa05818654b3816631540cf61d2414490520)). Dear editor @, if we want to add back a short literature review, do we need to remove trim other parts of the paper to keep the word counts?

> 2. Please provide a benchmark comparing the GP evaluation plus gradient computation time (or another suitable metric) to a comparable state-of-the-art model.

**Response**: We have added two benchmarks to the [py-demos](https://github.com/huanghua1994/HiGP/tree/main/py-demos) directory. A summary of experiment settings and results are available [here](https://github.com/huanghua1994/HiGP/blob/main/py-demos/benchmark-results.md).

> 3. l46f: for spatial 2D and 3D data, FFTs are usually an extremely efficient way to represent certain kernels see, e.g. KISS-GP. Since you mention these structures explicitly, please provide a brief description of the advantage that alternative representations yield compared to commonly used FFTs.

**Response**: We thank the reviewer for highlighting this important comparison. H-matrix methods offer distinct advantages over FFT-based approaches like KISS-GP or NFFT-based GP for spatial data:

* Non-stationary kernel support: H-matrices naturally accommodate spatially-varying covariance structures, essential for modeling heterogeneous phenomena. FFT methods are inherently limited to stationary kernels.

* Robustness for highly irregular data: For irregular datasets, H-matrices avoid the memory overhead of embedding data into regular grids required by KISS-GP. While NFFT-GP extends FFT methods to non-uniform grids, H-matrices excel when data exhibits extreme irregularity.

We note that future extensions of our framework could incorporate FFT-based fast matrix-vector multiplication methods for problems where they are most suitable, leveraging the complementary strengths of different approaches.

> 4. l52f: I'm not sure whether iterative methods plus preconditioners really set HiGP apart. AFAIK, many GP tools use iterative solvers with some kind of preconditioner including GPyTorch, the tool you cite. Even the ones that don't, tend to have a reason for not using one.

**Response**: Although many GP tools use iterative solver with preconditioner, the following contributions set HiGP apart, as we stated in the manuscript:

* Other GP tools, for example, GPyTorch, usually use a simple preconditioner that can be inefficient for some hyperparameter values. HiGP uses the AFN preconditioner, which is designed to be robust across a wide range of kernel hyperparameter values.

* HiGP uses hand-coded gradients for the preconditioner, which is more accurate and has better performance.

> 5. l58f: Iterative methods are usually not naively auto-diffed AFAIK (naively as in iterations being unrolled)! For example, JAX defines custom derivates for most iterative solvers and utilizes the implicit function theorem for retrieving gradients. Furthermore, using ``naked'' AD would not be safe as at least some Krylov subspace methods are prone to diverging gradients. IMO, you should focus on the gain in performance (with benchmarks) of having hand-coded gradients, not the efficiency of AD for iterative solvers.

**Response**: We aware that JAX defines custom derivates for some iterative solvers. The key issue is that the gradient of the preconditioner also needs to be computed, while some GP tools rely on AD for computing the graident of the preconditioner, which can be inefficient and inaccurate.

> 6. Could you please elaborate on the nature of the shared authorship and the contributions of each author?

**Response**: We have a private repository that houses early HiGP development code, including some experimental MATLAB code. When we released HiGP, we copied the C++ and Python code to the current public repository, and subsequent development will continue in the public repository. Contributions of each author:

* Hua Huang and Tianshi Xu: program development, documentation, paper writing
* Yuanzhe Xi and Edmond Chow: documentation and paper writing
