### 2025-09-11 58c2c5f
Add "# References" section
