### 2025-10-10 a5063a8
Paper revision based on the 2nd round review
