+++ paper/paper.bib
@@ -316,3 +316,75 @@
 	pages       = {1498-1504},
 	doi         = {10.1109/CDC49753.2023.10384051}
 }
+
+@article{Pinder:2022,
+  title   = {GPJax: A Gaussian Process Framework in JAX},
+  author  = {Pinder, Thomas and Dodd, Daniel},
+  journal = {Journal of Open Source Software},
+  volume  = {7},
+  number  = {75},
+  pages   = {4455},
+  year    = {2022},
+  doi     = {10.21105/joss.04455}
+}
+
+@article{Charlier:2021,
+  title   = {Kernel Operations on the GPU, with Autodiff, without Memory Overflows},
+  author  = {Charlier, Benjamin and Feydy, Jean and Glaunès, Joan Alexis and Collin, François-David and Durif, Ghislain},
+  journal = {Journal of Machine Learning Research},
+  volume  = {22},
+  number  = {74},
+  pages   = {1--6},
+  year    = {2021}
+}
+
+@article{Fairbrother:2022,
+  title   = {GaussianProcesses.jl: A Nonparametric Bayes Package for the Julia Language},
+  author  = {Fairbrother, Jamie and Nemeth, Christopher and Rischard, Maxime and Brea, Johanni and Pinder, Thomas},
+  journal = {Journal of Statistical Software},
+  volume  = {102},
+  number  = {1},
+  year    = {2022},
+  doi     = {10.18637/jss.v102.i01}
+}
+
+@inproceedings{Wenger:2022,
+  title     = {Preconditioning for Scalable Gaussian Process Hyperparameter Optimization},
+  author    = {Wenger, Jonathan and Pleiss, Geoff and Hennig, Philipp and Cunningham, John P. and Gardner, Jacob R.},
+  booktitle = {Proceedings of the 39th International Conference on Machine Learning},
+  year      = {2022},
+  pages     = {23751--23780},
+  publisher = {PMLR}
+}
+
+@inproceedings{Wu:2024,
+  title     = {Large-Scale Gaussian Processes via Alternating Projection},
+  author    = {Wu, Kenan and Meinshausen, Nicolai and Cunningham, John P. and Pozzi, Matteo},
+  booktitle = {Proceedings of the 27th International Conference on Artificial Intelligence and Statistics},
+  year      = {2024},
+  pages     = {2810--2818},
+  publisher = {PMLR}
+}
+
+@article{Pan:2024,
+  title   = {GPU-Accelerated Vecchia Approximations of Gaussian Processes for Geospatial Data using Batched Matrix Computations},
+  author  = {Pan, Qilong and Abdulah, Sameh and Genton, Marc G. and Keyes, David E. and Ltaief, Hatem and Sun, Ying},
+  journal = {arXiv preprint arXiv:2403.07412},
+  year    = {2024}
+}
+
+@inproceedings{Ament:2024,
+  title     = {Robust Gaussian Processes via Relevance Pursuit},
+  author    = {Sebastian Ament and Elizabeth Santorella and David Eriksson and Benjamin Letham and Maximilian Balandat and Eytan Bakshy},
+  booktitle = {The Thirty-eighth Annual Conference on Neural Information Processing Systems},
+  year      = {2024},
+  url       = {https://openreview.net/forum?id=5FATPIlWUJ}
+}
+
+@inproceedings{Lin:2025,
+  title     = {Scalable Gaussian Processes with Latent Kronecker Structure},
+  author    = {Jihao Andreas Lin and Sebastian Ament and Maximilian Balandat and David Eriksson and Jos{\'e} Miguel Hern{\'a}ndez-Lobato and Eytan Bakshy},
[…]
