### 2025-10-10 a11a28f
Update environment requirements
