### 2025-10-15 #8621 @huanghua1994
@jGaboardi @Edenhofer We will add the benchmarking content into documentation and reference that in the manuscript. Much appreciated for your understanding and suggestion.
