### 2025-10-15 52de716
Add performance test results to doc
