### 2025-10-15 b51d717
Add reference to performance test section in doc
