### 2025-10-21 #8621 @jGaboardi
@huanghua1994 Are you ready for the reviewers to make another (possibly final) pass to check that everything is in order?

cc @Edenhofer, @davidogara, @Naeemkh
