+++ docs/0-List-of-contents.md
@@ -1,6 +1,6 @@
 # HiGP Users Manual
 
-Gaussian processes (GPs) are a set of flexible, non-parametric Bayesian methods for modeling complex data. HiGP is a high-performance Python package for GP regression (GPR) and GP classification (GPC). This HiGP user manual contains four parts:
+Gaussian processes (GPs) are a set of flexible, non-parametric Bayesian methods for modeling complex data. HiGP is a high-performance Python package for GP regression (GPR) and GP classification (GPC). This HiGP user manual contains the following parts:
 
 1. [Basic usage of HiGP](https://github.com/huanghua1994/HiGP/blob/main/docs/1-Basic-usage-of-HiGP.md)
     * 1.1 Install HiGP
@@ -20,3 +20,8 @@ Gaussian processes (GPs) are a set of flexible, non-parametric Bayesian methods 
     * 4.4 Building the Python packages for release
     * 4.5 Debugging the C/C++ source code on local machine
 5. [Performance tests](https://github.com/huanghua1994/HiGP/blob/main/docs/5-Performance-tests.md)
+6. [Contribution Guidelines](https://github.com/huanghua1994/HiGP/blob/main/docs/6-Contribution-Guidelines.md)
+    * 6.1 How to Contribute
+    * 6.2 Filing Issues
+    * 6.3 Getting Support
+    * 6.4 License
