### 2025-11-02 huanghua1994/HiGP#5 @huanghua1994
Thank you for the suggestion. We have added a [Contribution Guidelines](https://github.com/huanghua1994/HiGP/blob/main/docs/6-Contribution-Guidelines.md) section in the documentation.
