### 2025-11-02 a6c4552
Add a single sentence summary of performance test results
as one reviewer required.
