### 2025-11-02 #8621 @huanghua1994
Hi @Naeemkh , as you required, I added a single sentence summary of the test results.

@jGaboardi Most of the issues in the repo are addressed and closed. The remaining one https://github.com/huanghua1994/HiGP/issues/3 requires an update to the `pyproject.toml`. We plan to move the package build method from the current `setup.py` to `pyproject.toml` later, and will update the package description. I think this should not be a blocker to the paper review.
