### 2025-11-02 #8621 @jGaboardi
@huanghua1994

> The remaining one huanghua1994/HiGP#3 requires an update to the `pyproject.toml`. We plan to move the package build method from the current `setup.py` to `pyproject.toml` later, and will update the package description. I think this should not be a blocker to the paper review.

Whether or not that's a blocker is determined by @Edenhofer who made the issue. Keep in mind that while the manuscript is absolutely a key component, [the core JOSS review process is designed to improve the quality of the software submitted](https://joss.theoj.org/about#:~:text=a%20formal%20peer%20review%20process%20that%20is%20designed%20to%20improve%20the%20quality%20of%20the%20software%20submitted). To that end, I would encourage you to revisit the reviewers' checklists[^1][^2][^3] and confirm with them what they believe is not yet addressed.

I'll also clarify [one of my comments](https://github.com/openjournals/joss-reviews/issues/8621#issuecomment-3424237077) because I feel I may have misrepresented my meaning. When I mentioned "*... (possibly final) pass*..." that was merely a colloquial statement – I didn't intend for that to mean that I thought all reviewer concerns had been addressed or that acceptance was imminent. My apologies if that caused any confusion.

Specifically regarding the benchmarking aspect, while I do not think the full inclusion of benchmarking is necessary in the manuscript itself (due to the JOSS [length limit](https://joss.readthedocs.io/en/latest/paper.html#:~:text=The%20paper%20should%20be%20between%20250%2D1000%20words.%20Authors%20submitting%20papers%20significantly%20longer%20than%201000%20words%20may%20be%20asked%20to%20reduce%20the%20length%20of%20their%20paper.) and [this bullet](https://joss.readthedocs.io/en/latest/paper.html#:~:text=A%20summary%20describing%20the%20high%2Dlevel%20functionality%20and%20purpose%20of%20the%20software%20for%20a%20diverse%2C%20non%2Dspecialist%20audience.)), I fully agree with @Naeemkh's [comment](https://github.com/openjournals/joss-reviews/issues/8621#issuecomment-3478087087) that the benchmarking details:

1. must be more visible within the repository structure (perhaps directly as a section in the `README.md`?),
2. should be tied to a specific version of `higp`, and
3. should be referenced through a full SHA-1 commit hash to an exact point-in-time, not linked to a particular branch like it is currently.

However, I have no problem being overruled by the Managing EiC (@danielskatz) in this matter if he thinks this is not in the best spirit of JOSS.

[^1]: @Naeemkh's [checklist](https://github.com/openjournals/joss-reviews/issues/8621#issuecomment-3094730384)
[^2]: @Edenhofer's [checklist](https://github.com/openjournals/joss-reviews/issues/8621#issuecomment-3105461141)
[^3]: @davidogara's [checklist](https://github.com/openjournals/joss-reviews/issues/8621#issuecomment-3172982076)
