### 2025-11-02 huanghua1994/HiGP#2 @Edenhofer
I think you have of documentation that warrants hosting on a dedicated website :)
