### 2025-11-02 f2acfce
Modifications following reviewers' comments
