### 2025-11-03 61b7ed6
Add python package desc
