### 2025-11-08 #8621 @jGaboardi
After conferring further with the editorial board and giving this more thought, I have some ideas on how we can move forward that *may* address reviewers concerns. Before we get into the "how" of it, we need to come to consensus on the "what" of it.

@Edenhofer, @davidogara, @Naeemkh - What performance & benchmarking information is needed for each of you to feel comfortable properly *assessing* the "high performance" claims of `higp`? Please be as detailed as needed/possible.

Once each of you has provided that information, we can determine the "how" of including that information in the manuscript without adding to the total length.

@huanghua1994 – While this is somewhat contrary to [my comment above](https://github.com/openjournals/joss-reviews/issues/8621#issuecomment-3399813063), I believe it will provide the reviewers the requisite evidence to appropriately make further decisions on this submission.
