### 2025-11-17 #8621 @Edenhofer
I'm a bit puzzled by the ping for further information. I believe performance comparisons are quite standard. I think any simple comparison that demonstrates that HiGP achieves better speed at the same quality would demonstrate its advantage well. Really any comparison to that end is fine with me.

For example, showing the achieved accuracy vs wall clock time would work well. Given that most of the heavy lifting is done outside of python, I'd be especially curious about large scale problems as python's overhead will surely dominante for small problems. The final results can be as short as saying, "we achieve accuracy of X equivalent to GPytorch at XX% of the runtime with 1 thread respectively ... for N threads on problem X."
