### 2025-11-18 #8621 @davidogara
> [@Edenhofer](https://github.com/Edenhofer), [@davidogara](https://github.com/davidogara), [@Naeemkh](https://github.com/Naeemkh)
>
> Gentle ping to ensure [my previous message](https://github.com/openjournals/joss-reviews/issues/8621#issuecomment-3506785340) didn't get lost in the shuffle. Specifically the following:
>
> > What performance & benchmarking information is needed for each of you to feel comfortable properly assessing the "high performance" claims of `higp`? Please be as detailed as needed/possible.

@jGaboardi thanks for the ping.

I think my remaining concerns would be addressed if the table that appears at the bottom of the [benchmark page](https://github.com/huanghua1994/HiGP/blob/main/docs/5-Performance-tests.md) was included in the manuscript. I am interested to see what the other reviewers think, but that looks like a perfectly reasonable performance test (and underlines the main contribution of the package).
