### 2025-11-30 #8621 @jGaboardi
* @davidogara
  * Thanks your comment above[^1].
* @Naeemkh
  * Do have any thoughts on @davidogara's suggestion or either of @huanghua1994 's comments above[^2][^3]?
* @huanghua1994
  * Let's wait another few days to see if there further comments from @Naeemkh and @Edenhofer, and then proceed with including that benchmark table as suggested in[^1][^3] .
  * Don't worry about word count *for the moment*; we can trim that down once the reviewers are satisfied with the content (or not).

[^1]: https://github.com/openjournals/joss-reviews/issues/8621#issuecomment-3544644163
[^2]: https://github.com/openjournals/joss-reviews/issues/8621#issuecomment-3559178220
[^3]: https://github.com/openjournals/joss-reviews/issues/8621#issuecomment-3559220885
