### 2025-12-05 #8621 @Naeemkh
Hi @jGaboardi,

I share @Edenhofer 's concern, and I find myself aligned with his assessment. As I mentioned in my [previous comment](https://github.com/openjournals/joss-reviews/issues/8621#issuecomment-3478087087), without more proactive engagement from the authors, the review process is becoming unnecessarily slow.

Regarding the reporting of scalability, the statement “HiGP is 539% to 8061% faster than GPyTorch” does not meaningfully inform the reader about the high-performance capabilities of the package. Since you appear to have access to a 24-core system, it would be far more informative to run consistent experiments across 1, 2, 4, 8, and 16 cores to demonstrate scalability and efficiency. For HPC-oriented work, reporting strong and weak scaling results is much more relevant than comparing performance to another package under potentially different conditions. In HPC work, it is generally expected that the authors themselves define appropriate scaling experiments that best demonstrate the performance characteristics of their system. The examples I mentioned are meant only to illustrate the type of analysis typically expected. Many HPC clusters offer nodes with 96+ cores, and readers would benefit from understanding how the package scales across increasing resources.

In response to my earlier feedback, the authors added a commit hash. I would not recommend including specific commit hashes in the manuscript; the paper should remain self-contained. However, If a reference to a particular version is absolutely necessary, using a tagged release rather than a raw commit hash would be more appropriate, as locating commits by hash is impractical for readers. While the reviewers can certainly offer detailed feedback when possible, addressing such fine-grained issues one by one is not an effective use of time and will inevitably prolong the review process.

These issues exemplify a broader lack of proactive effort from the authors. Some additional initiative would greatly streamline the review process. The package itself represents a substantial scientific contribution, but the way it is currently presented in the manuscript does not fully convey its value. A major revision would help the authors more clearly articulate the contribution and the message they aim to communicate to the community. Such revisions may be handled more effectively as a new submission.
