+++ paper/paper.md
@@ -75,10 +75,6 @@ Secondly, HiGP uses iterative solvers with the newly proposed AFN preconditioner
 
 Lastly, HiGP uses accurate and efficient hand-coded gradient calculations. GPyTorch relies on the automatic differentiation (autodiff) provided in PyTorch to calculate gradients (\autoref{eq:trace}). However, autodiff can be inefficient and inaccurate for computing the gradient of the preconditioner, so we use hand-coded gradient calculations for better performance and accuracy.
 
-The HiGP documentation[^1] provides a comparison of the accuracy and performance of HiGP version 2025.8.21 (git commit 8942631) and GPyTorch version 1.14. The tests were performed on one node (shared memory) with a 24-core 3.0 GHz Intel Xeon Gold 6248R CPU, using the "Bike Sharing" and "3D Road Network" data sets from the UCI Machine Learning Datasets and three synthetic target functions "Rosenbrock", "Rastrigin", and "Branin" from the Virtual Library of Simulation Experiments; HiGP is 539\% to 8061\% faster than GPyTorch.
-
-[^1]: https://github.com/huanghua1994/HiGP/blob/main/docs/5-Performance-tests.md
-
 # Design and Implementation
 
 We implemented HiGP in Python 3 and C++ with the goal of providing both a set of ready-to-use out-of-the-box Python interfaces for regular users and a set of reusable high-performance shared-memory multithreading computational primitives for advanced users. The HiGP C++ code implements all performance-critical operations. The HiGP Python code wraps the C++ units into four basic Python modules: `krnlmatmodule` for computing kernel matrices and its derivatives, `precondmodule` for PCG solver with AFN preconditioner, `gprproblemmodule` and `gpcproblemmodule` for computing the the loss and gradient for GP regression and classification. The two modules `gprproblemmodule` and `gpcproblemmodule` allow a user to train a GP model with any gradient-based optimizer.
@@ -100,5 +96,65 @@ pred = higp.gpr_prediction(train_x, train_y, test_x,
 ```
 
 We note that the HiGP Python interfaces (except for `GPRModel` and `GPCModel` models) are *stateless*. This design aims to simplify the interface and decouple different operations. A user can train and use different GP models with the same or different data and configurations in the same file.
+
+# Numerical Experiments
+
+We conducted numerical experiments on an Ubuntu 20.04 LTS machine with dual Intel Xeon Gold 6248R CPU (2x12 cores in total). We used PyTorch 2.8.0, GPyTorch 1.14, and HiGP version 2025.11.3 for the tests.
+
+We tested two data sets from the UCI Machine Learning Datasets: the "Bike Sharing" and the "3D Road Network" data sets. We also tested three synthetic target functions from the Virtual Library of Simulation Experiments with randomly sampled data points: Rosenbrock, Rastrigin, and Branin. All datasets were normalized with Z-score normalization ($\mu=0$, $\sigma=1$) applied to both features and targets using statistics from the training set. The results represent averages over three independent runs for statistical reliability. Both HiGP and GPyTorch were configured with identical computational budgets to ensure a fair comparison. E2E tests use the following settings:
+
+- Optimizer steps: 50
+- CG iterations: 20 (training), 50 (prediction)
[…]
