### 2025-12-17 da44859
Update performance test results
Previous results were obtained without correctly setting OpenMP binding
