### 2025-12-17 #8621 @huanghua1994
Dear editor @jGaboardi and reviewers @Naeemkh @davidogara ,

We have carefully reviewed **all** comments in this thread and made changes to the manuscript following all reviewers' comments. Specifically:

 ======== Comments from @davidogara ========

> I think my remaining concerns would be addressed if the table that appears at the bottom of the [benchmark page](https://github.com/huanghua1994/HiGP/blob/main/docs/5-Performance-tests.md) was included in the manuscript.

We have added that table to the manuscript.

======== Comments from @Naeemkh ========

> Since you appear to have access to a 24-core system, it would be far more informative to run consistent experiments across 1, 2, 4, 8, and 16 cores to demonstrate scalability and efficiency.

We have added one strong scaling test for only HiGP using 1, 2, 4, 8, and 16 cores on the same 24-core machine we used before.

> I would not recommend including specific commit hashes in the manuscript; the paper should remain self-contained. However, if a reference to a particular version is absolutely necessary, using a tagged release rather than a raw commit hash would be more appropriate, as locating commits by hash is impractical for readers.

We have added a tagged release and refer to the tag in the manuscript.

========== End of reviewer comments ===========

Much appreciated for your time and help!
