### 2026-01-01 #8621 @jGaboardi
Also, @davidogara – if you are satisfied with the state of the paper other than these last table clean-ups, etc, can you complete [your checklist above](https://github.com/openjournals/joss-reviews/issues/8621#issuecomment-3703088629)?
