### 2026-01-02 619804e
Update table caption and some paper DOIs
