+++ paper/paper.bib
@@ -11,13 +11,15 @@
 	title        = {Probabilistic Machine Learning: An introduction},
 	author       = {Murphy, Kevin P.},
 	year         = 2022,
-	publisher    = {MIT Press}
+	publisher    = {MIT Press},
+	isbn         = {978-0-262-04682-4}
 }
 @book{Murphy:2023,
 	title        = {Probabilistic Machine Learning: Advanced Topics},
 	author       = {Murphy, Kevin P.},
 	year         = 2023,
-	publisher    = {MIT Press}
+	publisher    = {MIT Press},
+	isbn         = {978-0-262-04843-9}
 }
 @inproceedings{Michalis:2009,
 	title        = {Variational Learning of Inducing Variables in Sparse {Gaussian} Processes},
@@ -38,7 +40,7 @@
 	publisher    = {AUAI Press},
     pages        = {282--290},
     location     = {Bellevue, WA},
-	url          = {https://dl.acm.org/doi/10.5555/3023638.3023667}
+	doi          = {10.5555/3023638.3023667}
 }
 @misc{Wilson:2015,
 	title        = {Thoughts on Massively Scalable {Gaussian} Processes},
@@ -104,7 +106,8 @@
 	publisher    = {Springer},
 	address      = {Berlin},
 	pages        = {9--29},
-	institution  = {{Max Planck Institute for Mathematics in the Sciences}}
+	institution  = {{Max Planck Institute for Mathematics in the Sciences}},
+	doi          = {10.1007/978-3-642-59709-1_2}
 }
 @article{Hackbusch:2002,
 	title        = {Data-sparse approximation by adaptive {$\mathcal{H}^2$-matrices}},
@@ -113,7 +116,8 @@
 	journal      = {Computing},
 	volume       = 69,
 	number       = 1,
-	pages        = {1--35}
+	pages        = {1--35},
+	doi          = {10.1007/s00607-002-1450-4}
 }
 @article{Xing:2019,
 	title        = {Interpolative decomposition via proxy points for kernel matrices},
@@ -132,7 +136,8 @@
 	journal      = {ACM Transactions on Mathematical Software},
 	volume       = 47,
 	number       = 1,
-	pages        = {1--29}
+	pages        = {1--29},
+	doi          = {10.1145/3412850}
 }
 @article{Cai:2023,
 	title        = {Data-driven Construction of Hierarchical Matrices with Nested Bases},
@@ -161,7 +166,8 @@
 	author       = {Gardner, Jacob and Pleiss, Geoff and Weinberger, Kilian Q and Bindel, David and Wilson, Andrew G},
 	year         = 2018,
 	journal      = {Advances in Neural Information Processing Systems},
-	volume       = 31
+	volume       = 31,
+	doi          = {10.5555/3327757.3327857}
 }
 @article{GPflow:2017,
 	title        = {{GPflow}: A {Gaussian} process library using {TensorFlow}},
@@ -171,7 +177,8 @@
 	journal      = {Journal of Machine Learning Research},
 	volume       = 18,
 	number       = 40,
-	pages        = {1--6}
+	pages        = {1--6},
+	doi          = {10.5555/3122009.3122049}
 }
 @misc{GPflow:2020,
 	title        = {A Framework for Interdomain and Multioutput {Gaussian} Processes},
@@ -228,7 +235,7 @@
 	series       = {Proceedings of Machine Learning Research},
 	volume       = 48,
 	pages        = {1472--1481},
-	editor       = {Balcan, Maria Florina and Weinberger, Kilian Q.}
+	doi          = {10.5555/3045390.3045546}
 }
[…]
