+++ paper/paper.bib
@@ -40,7 +40,7 @@
 	publisher    = {AUAI Press},
     pages        = {282--290},
     location     = {Bellevue, WA},
-	doi          = {10.5555/3023638.3023667}
+	url          = {https://dl.acm.org/doi/10.5555/3023638.3023667}
 }
 @misc{Wilson:2015,
 	title        = {Thoughts on Massively Scalable {Gaussian} Processes},
@@ -167,7 +167,7 @@
 	year         = 2018,
 	journal      = {Advances in Neural Information Processing Systems},
 	volume       = 31,
-	doi          = {10.5555/3327757.3327857}
+	url          = {https://dl.acm.org/doi/10.5555/3327757.3327857}
 }
 @article{GPflow:2017,
 	title        = {{GPflow}: A {Gaussian} process library using {TensorFlow}},
@@ -178,7 +178,7 @@
 	volume       = 18,
 	number       = 40,
 	pages        = {1--6},
-	doi          = {10.5555/3122009.3122049}
+	url          = {https://dl.acm.org/doi/10.5555/3122009.3122049}
 }
 @misc{GPflow:2020,
 	title        = {A Framework for Interdomain and Multioutput {Gaussian} Processes},
@@ -235,7 +235,7 @@
 	series       = {Proceedings of Machine Learning Research},
 	volume       = 48,
 	pages        = {1472--1481},
-	doi          = {10.5555/3045390.3045546}
+	url          = {https://dl.acm.org/doi/10.5555/3045390.3045546}
 }
 @inproceedings{Damianou:2013,
 	title        = {Deep {G}aussian Processes},
@@ -343,7 +343,7 @@
   number  = {74},
   pages   = {1--6},
   year    = {2021},
-  doi     = {10.5555/3546258.3546332}
+  url     = {https://dl.acm.org/doi/10.5555/3546258.3546332}
 }
 
 @article{Fairbrother:2022,
@@ -388,7 +388,7 @@
   author    = {Sebastian Ament and Elizabeth Santorella and David Eriksson and Benjamin Letham and Maximilian Balandat and Eytan Bakshy},
   booktitle = {The Thirty-eighth Annual Conference on Neural Information Processing Systems},
   year      = {2024},
-  doi       = {10.52202/079017-1972}
+  url       = {https://dl.acm.org/doi/10.5555/3737916.3739888}
 }
 
 @inproceedings{Lin:2025,
