### 2026-01-08 #8621 @jGaboardi
> One thing I should note here is that the Zenodo archive page shows my affiliation as NVIDIA since I am working at NVIDIA now and my GitHub account has linked with NVIDIA, while most of my work in this paper is completed before graduation, so my affiliation in the paper is my school instead of NVIDIA.

@huanghua1994, I think that is not a problem, but I'll tag the track EiC (@danielskatz) just for good measure while I work on my checklist.
