### 2026-01-08 huanghua1994/HiGP#13 @jGaboardi
[JOSS] Meyer2021 reference need update
* missing an `@` for the in the Meyer2021 reference at the end of line 66  – line 39 in the PDF
* xref https://github.com/openjournals/joss-reviews/issues/8621

https://github.com/huanghua1994/HiGP/blob/bad5c2fb9adb2e05df878096118416a1e7b8e77a/paper/paper.md?plain=1#L66

@huanghua1994
