### 2026-01-08 #8621 @jGaboardi
After you address https://github.com/huanghua1994/HiGP/issues/13 and my [comment above](https://github.com/openjournals/joss-reviews/issues/8621#issuecomment-3721457977), ping me back and I will continue my process.
