### 2026-01-09 huanghua1994/HiGP#14 @jGaboardi
[JOSS] duplicate `Wenger:2022` references
* There are duplicate references for `Wenger:2022` which are the same paper
  * https://github.com/huanghua1994/HiGP/blob/e9a54efcccb06062522567f97bbac4bd68c0e366/paper/paper.bib#L265
  * https://github.com/huanghua1994/HiGP/blob/e9a54efcccb06062522567f97bbac4bd68c0e366/paper/paper.bib#L359
* please remove one
* xref: https://github.com/openjournals/joss-reviews/issues/8621#issuecomment-3726674762
