### 2026-01-09 c917e32
Clean up bib file
