+++ paper/paper.bib
@@ -1,189 +1,130 @@
-% GP related
-@book{Rasmussen:2005,
-	title        = {Gaussian Processes for Machine Learning},
-	author       = {Rasmussen, Carl Edward and Williams, Christopher K. I.},
-	year         = 2005,
-	publisher    = {The MIT Press},
-	doi          = {10.7551/mitpress/3206.001.0001},
-	isbn         = {978-0-262-25683-4}
+@inproceedings{Ament:2024,
+	title        = {Robust Gaussian Processes via Relevance Pursuit},
+	author       = {Sebastian Ament and Elizabeth Santorella and David Eriksson and Benjamin Letham and Maximilian Balandat and Eytan Bakshy},
+	year         = {2024},
+	booktitle    = {The Thirty-eighth Annual Conference on Neural Information Processing Systems},
+	url          = {https://dl.acm.org/doi/10.5555/3737916.3739888}
 }
-@book{Murphy:2022,
-	title        = {Probabilistic Machine Learning: An introduction},
-	author       = {Murphy, Kevin P.},
-	year         = 2022,
-	publisher    = {MIT Press},
-	isbn         = {978-0-262-04682-4}
+@article{Aune:2014,
+	title        = {Parameter estimation in high dimensional {Gaussian} distributions},
+	author       = {Erlend Aune and Daniel P. Simpson and Jo Eidsvik},
+	year         = {2014},
+	journal      = {Statistics and Computing},
+	volume       = {24},
+	number       = {2},
+	pages        = {247--263},
+	doi          = {10.1007/s11222-012-9368-y}
 }
-@book{Murphy:2023,
-	title        = {Probabilistic Machine Learning: Advanced Topics},
-	author       = {Murphy, Kevin P.},
-	year         = 2023,
-	publisher    = {MIT Press},
-	isbn         = {978-0-262-04843-9}
-}
-@inproceedings{Michalis:2009,
-	title        = {Variational Learning of Inducing Variables in Sparse {Gaussian} Processes},
-	author       = {Michalis Titsias},
-	year         = 2009,
-	month        = apr,
-	booktitle    = {Proceedings of the Twelth International Conference on Artificial Intelligence and Statistics},
+@inproceedings{Balcan:2016,
+	title        = {Deep {Gaussian} Processes for Regression using Approximate Expectation Propagation},
+	author       = {Bui, Thang and Hernandez-Lobato, Daniel and Hernandez-Lobato, Jose and Li, Yingzhen and Turner, Richard},
+	year         = {2016},
+	month        = {20--22 Jun},
+	booktitle    = {Proceedings of The 33rd International Conference on Machine Learning},
 	publisher    = {PMLR},
-	volume       = 5,
-	pages        = {567--574},
-	editor       = {David van Dyk and Max Welling}
+	address      = {New York, New York, USA},
+	series       = {Proceedings of Machine Learning Research},
+	volume       = {48},
+	pages        = {1472--1481},
+	url          = {https://dl.acm.org/doi/10.5555/3045390.3045546}
 }
-@article{Hensman:2013,
-	title        = {{Gaussian} Processes for Big Data},
-	author       = {Hensman, James and Fusi, Nicol\`{o} and Lawrence, Neil D.},
-	year         = 2013,
-	month        = sep,
-	publisher    = {AUAI Press},
-    pages        = {282--290},
-    location     = {Bellevue, WA},
-	url          = {https://dl.acm.org/doi/10.5555/3023638.3023667}
-}
-@misc{Wilson:2015,
[…]
