### 2026-01-09 huanghua1994/HiGP#14 @huanghua1994
Thank you for pointing this out. We have cleaned up the bib file and checked in the changes.
