### 2026-01-09 474511b
Remove duplicate bib item
