+++ paper/paper.bib
@@ -340,14 +340,6 @@
 	publisher    = {PMLR},
 	pages        = {23751--23780}
 }
-@inproceedings{Wenger:2022,
-	title        = {Preconditioning for Scalable Gaussian Process Hyperparameter Optimization},
-	author       = {Wenger, Jonathan and Pleiss, Geoff and Hennig, Philipp and Cunningham, John P. and Gardner, Jacob R.},
-	year         = {2022},
-	booktitle    = {Proceedings of the 39th International Conference on Machine Learning},
-	publisher    = {PMLR},
-	pages        = {23751--23780}
-}
 @misc{Wilson:2015,
 	title        = {Thoughts on Massively Scalable {Gaussian} Processes},
 	author       = {Wilson, Andrew Gordon and Dann, Christoph and Nickisch, Hannes},
