### 2026-01-09 c269a06
Fix capitalization in titles in the bib file
