+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @inproceedings{Ament:2024,
-	title        = {Robust Gaussian Processes via Relevance Pursuit},
+	title        = {Robust {Gaussian} Processes via Relevance Pursuit},
 	author       = {Sebastian Ament and Elizabeth Santorella and David Eriksson and Benjamin Letham and Maximilian Balandat and Eytan Bakshy},
 	year         = {2024},
 	booktitle    = {The Thirty-eighth Annual Conference on Neural Information Processing Systems},
@@ -63,7 +63,7 @@
 	doi          = {10.1109/CDC49753.2023.10384051}
 }
 @inproceedings{Damianou:2013,
-	title        = {Deep {G}aussian Processes},
+	title        = {Deep {Gaussian} Processes},
 	author       = {Damianou, Andreas and Lawrence, Neil D.},
 	year         = {2013},
 	month        = {29 Apr--01 May},
@@ -93,7 +93,7 @@
 	doi          = {10.1080/10618600.2024.2353653}
 }
 @article{Fairbrother:2022,
-	title        = {GaussianProcesses.jl: A Nonparametric Bayes Package for the Julia Language},
+	title        = {GaussianProcesses.jl: A Nonparametric Bayes Package for the {Julia} Language},
 	author       = {Fairbrother, Jamie and Nemeth, Christopher and Rischard, Maxime and Brea, Johanni and Pinder, Thomas},
 	year         = {2022},
 	journal      = {Journal of Statistical Software},
@@ -128,7 +128,7 @@
 	url          = {https://arxiv.org/abs/2003.01115}
 }
 @misc{GPy:2014,
-	title        = {{GPy}: A {Gaussian} process framework in Python},
+	title        = {{GPy}: A {Gaussian} process framework in {Python}},
 	author       = {{GPy}},
 	year         = {since 2012},
 	howpublished = {http://github.com/SheffieldML/GPy}
@@ -213,7 +213,7 @@
 	url          = {https://arxiv.org/abs/2106.12135}
 }
 @inproceedings{Lin:2025,
-	title        = {Scalable Gaussian Processes with Latent Kronecker Structure},
+	title        = {Scalable {Gaussian} Processes with Latent {Kronecker} Structure},
 	author       = {Jihao Andreas Lin and Sebastian Ament and Maximilian Balandat and David Eriksson and Jos{\'e} Miguel Hern{\'a}ndez-Lobato and Eytan Bakshy},
 	year         = {2025},
 	booktitle    = {Forty-second International Conference on Machine Learning},
@@ -254,7 +254,7 @@
 	isbn         = {978-0-262-04843-9}
 }
 @article{Pan:2024,
-	title        = {GPU-Accelerated Vecchia Approximations of Gaussian Processes for Geospatial Data using Batched Matrix Computations},
+	title        = {{GPU}-Accelerated Vecchia Approximations of {Gaussian} Processes for Geospatial Data using Batched Matrix Computations},
 	author       = {Pan, Qilong and Abdulah, Sameh and Genton, Marc G. and Keyes, David E. and Ltaief, Hatem and Sun, Ying},
 	year         = {2024},
 	booktitle    = {ISC High Performance 2024 Research Paper Proceedings (39th International Conference)},
@@ -262,7 +262,7 @@
 	doi          = {10.23919/ISC.2024.10528930}
 }
 @article{Pinder:2022,
-	title        = {GPJax: A Gaussian Process Framework in JAX},
+	title        = {{GPJax}: A {Gaussian} Process Framework in {JAX}},
 	author       = {Pinder, Thomas and Dodd, Daniel},
 	year         = {2022},
[…]
