### 2026-01-09 a4dde50
Fix capitalization in titles in the bib file
