+++ paper/paper.bib
@@ -179,7 +179,7 @@
 	url          = {https://dl.acm.org/doi/10.5555/3023638.3023667}
 }
 @article{Huang:2020,
-	title        = {{H2Pack}: High-performance \textit{{H}} $^{\textrm{2}}$ Matrix Package for Kernel Matrices Using the Proxy Point Method},
+	title        = {{H2Pack}: High-performance $\mathcal{H}^2$ Matrix Package for Kernel Matrices Using the Proxy Point Method},
 	author       = {Huang, Hua and Xing, Xin and Chow, Edmond},
 	year         = {2020},
 	month        = {12},
@@ -196,7 +196,7 @@
 	url          = {https://arxiv.org/abs/2305.07546}
 }
 @article{Hutchinson:1989,
-	title        = {A Stochastic Estimator of the Trace of the Influence Matrix for Laplacian Smoothing Splines},
+	title        = {A Stochastic Estimator of the Trace of the Influence Matrix for {Laplacian} Smoothing Splines},
 	author       = {Hutchinson, M.F.},
 	year         = {1989},
 	month        = jan,
