### 2026-01-09 #8621 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
